% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concentrations_unweighted_exposures}
\alias{concentrations_unweighted_exposures}
\title{Concentrations Exposure Data}
\format{
A data frame with 10123 rows and 10 variables:
\describe{
  \item{\code{byvar}}{character }
  \item{\code{LowerCI}}{double }
  \item{\code{UpperCI}}{double }
  \item{\code{Mean}}{double }
  \item{\code{Median}}{double }
  \item{\code{LowerPI}}{double }
  \item{\code{UpperPI}}{double }
  \item{\code{Year}}{character }
  \item{\code{UrbanRural}}{character }
  \item{\code{name}}{character }
}
}
\usage{
concentrations_unweighted_exposures
}
\description{
Contains yearly pm2.5 measurements for countries and regions
}
\details{
For further details, see \url{https://www.who.int/data/gho/data/themes/air-pollution/who-modelled-estimates-of-air-pollution-from-particulate-matter}
}
\keyword{datasets}
