% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exceed}
\alias{exceed}
\title{exceed}
\format{
A data frame with seventeen variables:
\describe{
\item{\code{ID}}{Name of Category: Country, GDB Region, GDB Super Region, Global, SDG Region, WHO Income Region, WHO Region e.g. Algeria or Western Asia and Northern Africa}
\item{\code{UrbanRural}}{Urban, Rural or Overall land classifcation}
\item{\code{Category}}{Type of Category: Country, GDB Region, GDB Super Region, Global, SDG Region, WHO Income Region, WHO Region}
\item{\code{Year}}{According year}
\item{\code{Value}}{Level of exposure}
\item{\code{Scale}}{Exceedance scale}
}

For further details, see \url{https://www.who.int/data/gho/data/themes/air-pollution/who-modelled-estimates-of-air-pollution-from-particulate-matter}
}
\usage{
exceed
}
\description{
DIMAWQ model predictions provide estimates of annual concentrations to PM2.5 at 0.1° × 0.1° high spatial resolution, which equates to approximately 11x11km at the equator globally.
}
\keyword{datasets}
