% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ground_monitors}
\alias{ground_monitors}
\title{Ground monitor Data}
\format{
A data frame with 34877 rows and 31 variables:
\describe{
  \item{\code{ID}}{integer }
  \item{\code{StationID}}{character }
  \item{\code{ISO3}}{character }
  \item{\code{CountryName}}{character }
  \item{\code{CountryCode}}{double }
  \item{\code{Year}}{integer }
  \item{\code{PM25}}{double }
  \item{\code{PM10}}{double }
  \item{\code{PM25PercCoverage}}{double }
  \item{\code{PM10PercCoverage}}{double }
  \item{\code{PM25Grading}}{integer }
  \item{\code{PM10Grading}}{integer }
  \item{\code{Longitude}}{double }
  \item{\code{Latitude}}{double }
  \item{\code{MonitorType}}{character }
  \item{\code{UnspecifiedType}}{integer }
  \item{\code{ApproxLoc}}{integer }
  \item{\code{PM25Conv}}{integer }
  \item{\code{IDGRID}}{double }
  \item{\code{SAT}}{double }
  \item{\code{DUST}}{double }
  \item{\code{SANOC}}{double }
  \item{\code{POP}}{double }
  \item{\code{ELEVATION}}{double }
  \item{\code{ELEVDIFFALTD}}{double }
  \item{\code{GBDRegion}}{character }
  \item{\code{GBDSuperRegion}}{character }
  \item{\code{WHORegion}}{character }
  \item{\code{WHOIncomeRegion}}{character }
  \item{\code{PredMethod}}{character }
  \item{\code{ConvFactor}}{double }
}
}
\usage{
ground_monitors
}
\description{
Contains yearly measurements from global ground monitor sites
}
\details{
For further details, see \url{https://www.who.int/data/gho/data/themes/air-pollution/who-modelled-estimates-of-air-pollution-from-particulate-matter}
}
\keyword{datasets}
