% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_weighted_concentrations}
\alias{population_weighted_concentrations}
\title{Population Weighted Concentrations}
\format{
A data frame with 10383 rows and 11 variables:
\describe{
  \item{\code{byvar}}{character }
  \item{\code{LowerCI}}{double }
  \item{\code{UpperCI}}{double }
  \item{\code{Mean}}{double }
  \item{\code{Median}}{double }
  \item{\code{LowerPI}}{double }
  \item{\code{UpperPI}}{double }
  \item{\code{Year}}{character }
  \item{\code{UrbanRural}}{character }
  \item{\code{Type}}{character }
  \item{\code{Category}}{character }
}
}
\usage{
population_weighted_concentrations
}
\description{
Contains yearly pm2.5 concerntraions based on population weighting
}
\details{
For further details, see \url{https://www.who.int/data/gho/data/themes/air-pollution/who-modelled-estimates-of-air-pollution-from-particulate-matter}
}
\keyword{datasets}
