% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pred_2013}
\alias{pred_2013}
\title{pred_2013}
\format{

}
\usage{
pred_2013
}
\description{
DIMAWQ model predictions provide estimates of annual concentrations to PM2.5 at 0.1° × 0.1° high spatial resolution, which equates to approximately 11x11km at the equator globally.
}
\keyword{datasets}
