% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pred_all}
\alias{pred_all}
\title{Join yearly grid_prediction data}
\usage{
pred_all()
}
\value{
a dataframe with all yearly grid_prediction data
}
\description{
Join yearly grid_prediction data
}
\examples{
# Read all data
yearly_pred = pred_all()
}
