% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_world_map}
\alias{who_world_map}
\title{WHO World Map SF}
\format{
A data frame with 237 rows and 13 variables:
\describe{
  \item{\code{CNTRY_TERR}}{integer }
  \item{\code{SOVEREIGN}}{integer }
  \item{\code{ISO_3_CODE}}{integer }
  \item{\code{ISO_2_CODE}}{integer }
  \item{\code{UN_CODE}}{double }
  \item{\code{WHO_CODE}}{double }
  \item{\code{WHO_REGION}}{integer }
  \item{\code{WHO_STATUS}}{integer }
  \item{\code{LVLID}}{integer }
  \item{\code{Shape_Leng}}{double }
  \item{\code{Shape_Area}}{double }
  \item{\code{CountryCode}}{double }
  \item{\code{geometry}}{list }
}
}
\usage{
who_world_map
}
\description{
SF dataframe containing WHO World map boundaries
}
\details{
For further details, see \url{https://www.who.int/data/gho/data/themes/air-pollution/who-modelled-estimates-of-air-pollution-from-particulate-matter}
}
\keyword{datasets}
