% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exceed}
\alias{exceed}
\title{Global Exceedance Data}
\format{
A data frame with 23548 rows and 6 variables:
\describe{
  \item{\code{ID}}{character }
  \item{\code{UrbanRural}}{character }
  \item{\code{Category}}{character }
  \item{\code{Year}}{double }
  \item{\code{Value}}{double }
  \item{\code{Scale}}{double }
}
}
\usage{
exceed
}
\description{
Contains yearly pm2.5 exceedance at 10,15,25.30 scale
}
\details{
For further details, see \url{https://www.who.int/data/gho/data/themes/air-pollution/who-modelled-estimates-of-air-pollution-from-particulate-matter}
}
\keyword{datasets}
